Fix `Crm#set_rights_to_case` request
Split crm cases method to separate module
Split crm cases test to seprate spec
