+++ .rubocop_todo.yml
@@ -1,6 +1,6 @@
 # This configuration was generated by
 # `rubocop --auto-gen-config`
-# on 2019-04-23 12:30:13 +0300 using RuboCop version 0.67.2.
+# on 2019-04-23 15:05:37 +0300 using RuboCop version 0.67.2.
 # The point is for the user to remove these configuration records
 # one by one as the offenses are removed from the code base.
 # Note that changes in the inspected code, or installation of new
@@ -15,16 +15,16 @@ Lint/UriEscapeUnescape:
 Metrics/AbcSize:
   Max: 23
 
-# Offense count: 16
+# Offense count: 17
 # Configuration parameters: CountComments, ExcludedMethods.
 # ExcludedMethods: refine
 Metrics/BlockLength:
-  Max: 1248
+  Max: 1127
 
 # Offense count: 4
 # Configuration parameters: CountComments.
 Metrics/ClassLength:
-  Max: 589
+  Max: 541
 
 # Offense count: 1
 Metrics/CyclomaticComplexity:
@@ -54,6 +54,7 @@ Naming/AccessorMethodName:
     - 'lib/teamlab/modules/community/community_forums.rb'
     - 'lib/teamlab/modules/community/community_wiki.rb'
     - 'lib/teamlab/modules/crm.rb'
+    - 'lib/teamlab/modules/crm/crm_cases.rb'
     - 'lib/teamlab/modules/files.rb'
     - 'lib/teamlab/modules/group.rb'
     - 'lib/teamlab/modules/mail.rb'
